Refactor Drag n Drop component
- Separate hooks for board logic and drag handlers
